Fix motion border exclusion checker
Previously there was a discrepency between "motion_border"
exclusion label calls made between
ophys_etl.transform.roi_transforms._motion_exclusion and
allensdk.brain_observatory.roi_masks.RoiMask.init_by_mask.

During trace extraction, the AllenSDK will set traces to NaN
for ROIs that overlap the motion border determined by
RoiMask.init_by_mask.

Because of this mismatch, ROIs that the ophys_etl
pipeline considered 'valid' but that AllenSDK considered
'invalid' resulted in the ophys_etl classifier module
loading in traces full of NaNs.

This commit changes how ophys_etl determines the
"motion_border" exclusion label so that it is in line
with how the AllenSDK determines the "motion_border"
exclusion label.

+++ src/ophys_etl/types/_dense_roi.py
@@ -15,9 +15,9 @@ class DenseROI(TypedDict):
     height: int
     valid_roi: bool
     mask_matrix: List[List[bool]]
-    max_correction_up: int
-    max_correction_down: int
-    max_correction_left: int
-    max_correction_right: int
+    max_correction_up: float
+    max_correction_down: float
+    max_correction_left: float
+    max_correction_right: float
     mask_image_plane: int
     exclusion_labels: List[str]
